# Remove most shutdown latency