### Details
When bedrock shuts down, there are a series of points at which it waits on things like network activity. These tend to have a 1 second timeout.

The shutdown procedure works by sending a signal, which sets a flag, and then waits for each of these loops to notice. Generally it takes a second for one loop to notice, which then does something, and it takes another loop a second to notice, etc.

This means it generally takes about 4 seconds to shut bedrock down. This change mostly just makes these loops shorter, to 100ms, and now it takes about 0.4 seconds to shut bedrock down. in production, this hardly matters, but in tests, where we start and stop a thousand bedrock processes, the 3.5 seconds per test makes a real difference. Auth tests run about 20% faster with this change.

For whatever reason Travis hates this change.

### Fixed Issues
Fixes GH_LINK

### Tests

Improved:
```
vagrant@expensidev2004:/vagrant/Auth/test$ time ./authtest -threads 16

Three run results:
real	5m20.448s
real	5m47.631s
real	5m31.428s
```

Before:
```
real	6m56.730s
real	7m18.563s
real	6m31.615s
```

Approximately 20% faster to run tests.